+++ apps/web/CHANGELOG.md
@@ -1,5 +1,13 @@
 # web
 
+## 0.3.5
+
+### Patch Changes
+
+- added dropdown and listview
+- Updated dependencies
+  - @duck4i/retro-ui@0.3.5
+
 ## 0.3.4
 
 ### Patch Changes